# specs/001-code-learning-platform/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: AI Code Learning Platform API
  description: |
    REST API for the AI Code Learning Platform - Phase 1 MVP.

    This platform transforms AI-generated code into beginner-friendly learning materials
    through task-based code uploads, automatic analysis, document generation, and Q&A.

    Powered by Google Gemini 2.5 Flash for fast, cost-effective AI content generation.

    **Authentication**: JWT-based with HTTPOnly cookies
    **Base URL**: `/api/v1`
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.codelearning.com/api/v1
    description: Production server
...
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: List user's projects
      description: Get all active (non-trashed) projects for authenticated user
      security:
        - cookieAuth: []
      parameters:
        - name: include_trashed
          in: query
          description: Include trashed projects
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Projects retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
                  total:
                    type: integer
                    example: 5

    post:
      tags: [Projects]
      summary: Create new project
      description: Create a new learning project container
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 255
                  example: My First Calculator
                description:
                  type: string
                  nullable: true
                  example: Learning to build a calculator with AI help
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Invalid input
        '401':
          description: Not authenticated
  /projects/{project_id}:
    get:
      tags: [Projects]
      summary: Get project details
      description: Get single project with task summary
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Project retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetail'
        '404':
          description: Project not found

    patch:
      tags: [Projects]
      summary: Update project
      description: Update project title or description
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'

    delete:
      tags: [Projects]
      summary: Soft delete project
      description: Move project and all tasks to trash (30-day retention)
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '204':
          description: Project moved to trash
        '404':
          description: Project not found

  # ==================== Tasks ====================
...
  /trash/projects/{project_id}/restore:
    post:
      tags: [Trash]
      summary: Restore project from trash
      description: Restore project and all its tasks
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Project restored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: JWT access token in HTTPOnly cookie
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        last_activity_at:
          type: string
          format: date-time
        deletion_status:
          type: string
          enum: [active, trashed]
        trashed_at:
          type: string
          format: date-time
          nullable: true
        task_count:
          type: integer
          description: Total tasks in project
        completed_task_count:
          type: integer
          description: Completed tasks
        progress_percentage:
          type: integer
          description: Overall project progress (0-100)
    ProjectDetail:
      allOf:
        - $ref: '#/components/schemas/Project'
        - type: object
          properties:
            tasks:
              type: array
              items:
                $ref: '#/components/schemas/Task'
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        task_number:
          type: integer
          example: 1
        title:
          type: string
        description:
          type: string
          nullable: true
        upload_method:
          type: string
          enum: [file, folder, paste]
        created_at:
          type: string
          format: date-time
        deletion_status:
          type: string
          enum: [active, trashed]
        document_status:
          type: string
          enum: [pending, in_progress, completed, failed]
          description: Learning document generation status
...